requirements.txt
uvicorn==0.34.0
shiny>=1.1.0
dotenv
faicons
plotly
numpy
shinywidgets
pandas
google.genai
google-cloud-bigquery
google-generativeai
markdown2
openpyxl
chatlas
asyncio
tabulate
db_dtypes
langchain
xlsxwriter
crewAI==0.102.0
langchain-community
langchain-core
duckdb
langchain-google-vertexai
pydantic==2.8.2
kaleido